[raft/scd] Move opintent upsert helpers into actions
